+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
   - Optional Phel file filter using equals syntax: `--debug="core"` or `--debug="boot"`
 - Add `phel eval` command
 - Make `repl` the default command when no args are provided to the executable
+- Move Facade Interfaces to Shared Module
 
 ## [0.22.2](https://github.com/phel-lang/phel-lang/compare/v0.22.1...v0.22.2) - 2025-09-23